Revert "Refactor mechanics.update to more scalable pattern"
This change results in a net code size gain, a performance
regression, and also doesn't necessarily map well to the eventual
pattern we will want. The motivation was to make it easier to add
more optional arguments like log without bloating the param list
with anytypes, but actually storing them results in a bunch of
loads/worse inlining etc. Its possible much of the performance
overhead could be done away with if battle were still passed around
as a param and the options were the only thing saved in the
Mechanics struct, but ultimately the pattern which is going to map
better to the C/JS API is just to make the final `log: anytype`
param a `options: anytype` param with `log`, `chance`, calc`, etc
fields.

This reverts commit 5ea92492cf587a0c2765e95abba6637e7bded35f.