Organize all meaningful RNG calls into Rolls
Spurious PS RNG advances which are just required for accuracy and
don't actually meaningfully affect the probability of events remain
untouched.

This refactoring is done in preparation for supporting
"-Dchance", in order to make it easier to find all the places the
turn probability chance will need to be updated.