OPTIONS -> CHOICES
in your API, breaking all yo shit

planning to maybe use "options" to name the struct that will get
passed to the battle update function after -Dchance happens, but
"options" is already kind of overloaded because its also used for
build options. either way, CHOICES probably was a better choice for
this anyway, so we're renaming. soz, welcome to 0.x!