add actions eql check to properly handle moveslots
we can't know which moveSlots are valid ahead of time because we
could have an opponent with less than 4 valid move slots (none or 0
pp) switch into disable or mimic. creating an RNG action with an
invalid slot will still return something with the same shape, so we
need to also verify that it matches the exact action a we passed in.

Action equality can be verified very equickly (3 instructions - xor
xor or) but only if we also ensure the override *durations* match
(not just that theyre the same shape). Thus, we must pass what we
expect the chance action to become instead of just {0, 1} so that
the input overrides do actually exactly equal the output actions.

fsr the battle rngs are not equal in all cases where change actions
are equal, which seems incorrect as that would require that the
actions dont fully account for all RNG (and if they didn't we would
expect inequality). revist this