Revert "add shape change damage skip"
This reverts commit 5918e03b2d494f75491387dca3b1c5c07eb1719f.

Turns out this is *not* sounds because it isn't guaranteed that all
damage values have the same shape. We could have discovered a shape
change while iterating over our original shape but can't know when
the original shape happened. eg. say our original shape was derived
from damage 240 (and a shape change occurs at 230) we can't skip
after noticing 217 is a different shape (nor can we jump to 240)
or else we dont actually cover the 230->255.