rework gen2 type order for fast sandstorm checks
With the previous type ordering, checking for sandstormImmune by
comparing types.includes(.Rock) etc results in:

    movzx   ecx, byte ptr [esp + 4]
    movzx eax, byte ptr [esp + 8]
    mov   edx, ecx
    and   dl, -2
    cmp   dl, 4
    sete  dl
    cmp   cl, 9
    sete  cl
    cmp   al, 4
    sete  ah
    or    ah, cl
    or    ah, dl
    cmp   al, 5
    sete  cl
    cmp   al, 9
    sete  al
    or    al, cl
    or    al, ah
    ret

This doesn't *really* matter in gen 2 because sand is rather rare,
but in future gens this will matter as sand will happen every turn.
We can try to be clever by adding back a placeholder and rearranging
Ground/Rock/Steel at the end so they all have the 5th bit set,
allowing us to then mask Types with 0b00010000_00010000, but this
ends up not only having more instructions (below) but also wastes 55
bytes of space (37 for the matchup chart, 16 for hidden power array
and 2 more bytes for the other two tables):

    mov ecx, dword ptr [esp + 8]
    movzx eax, byte ptr [esp + 4]
    shl   ecx, 8
    or    ecx, eax
    test  ecx, 4112
    setne al
    ret

  Another idea is to rearrange the types so that only
  Ground/Rock/Steel/???? have 0b011xx set (doesnt matter if ??? is a
  false positive because no actual Types will contain it), but there
  doesnt appear to be a good way to actually leverage this, so while
  it doesn't add space overhead it still isn't most efficient:

    movzx   ecx, byte ptr [esp + 8]
    movzx eax, byte ptr [esp + 4]
    not   cl
    not   al
    test  cl, 12
    sete  cl
    test  al, 12
    sete  al
    or    al, cl
    ret

  The most efficient is simply to rearrange Ground/Rock/Steel to be
  at the start (so as not to screw up the special check) and simply
  do range checks on type1 and type2, as done in this commit:

    cmp   byte ptr [esp + 4], 13
    setae cl
    cmp   byte ptr [esp + 8], 13
    setae al
    or    al, cl
    ret

We do need to add an additional 16 bytes for the HIDDEN_POWER table,
but pretty much any reordering of types would require that.
Furthermore, this commit adds in the INDEXES required for the
future Present implementation (though could have been implemented
somewhat inefficiently before this reordering, so perhaps its only
fair to consider that as overhead as well).

ideas and help from @Marty-D and @pasyg (though neither are to blame
for any negative consequences this change might result in...)