add preliminary build.zig.zon
also adds support to the build.zig for dealing with cases where
tools/test files are missing, as we don't want to include this stuff
in the package to end users.

this does include the `.tmpl` files as well as support for building
the node and WASM plugins, but the latter at least is intentional as
they are meant to be exposed by `options.interal` to allow for
creating "uber" binaries.