allow for overriding high-level RNG functionality
This is distinct from -Dchance in that there are some things that
are cleaner/easier to express by overriding the RNG than can be done
via forcing certain outcomes (i.e. the approach used by
Athena/FSAI).

Unfortunately this is difficult to expose via the C API/JS so
non-Zig users will need to modify the bindings themselves for their
own use cases.